GeoEngine tile update cleanup
This commit consists of multiple overlapping changes all with the same
aim which is to add an extra layer of abstraction to the tile grid so
that we can cope with different sized tiles for our audio UI.

Features:
 * The upgrading of the tile grid is no longer done on a timer, but
  when the location moves out of the central grid region. This makes
  it much more up to date.
 * Display of tile grid outline and 'central area' region on map UI.
  This can be enabled from the settings and is to allow easier debug.
  The feature is only visible in debug builds. The grid shown is the
  one currently in use by the GeoEngine. Note that when changing the
  setting the service needs a manual restart (sleep/unsleep).
  The new GeoJSON layer in our maplibre UI shows that there's
  potential for other debug uses.
 * Support 2x2 grid in addition to 3x3 grid. The reason for this is
  that the protomaps vector tiles are at zoom level 15 and not zoom
  level 16. A 3x3 grid at zoom level 15 is more data than we need,
  so we can now switch to using a 2x2 grid. The central area of the
  grid is where the current location can be without resulting in a
  new grid being generated. Once the user strays outside it, then a
  new grid is required. Although this was only required for 2x2, the
  same logic is used for 3x3 grids too. The result is that there's
  now some hysteresis so that if a user is walking down a tile
  boundary the grid won't continually be updated as they stray back
  and forth from one side of the tile boundary to the other.

Minor tidying:
 * All zoom levels are now Int rather than Double.
 * The grid size and zoom levels for the TileGrid are now declared
  inside the TileGrid class. They could be private and const, except
  that the unit test uses them.
 * Moved getTileGrid functions into their own file and it now returns
  a TileGrid object which contains the central region as well as the
  list of tiles.

+++ app/src/main/java/org/scottishtecharmy/soundscape/utils/GeoUtils.kt
@@ -191,7 +191,6 @@ fun pixelXYToLatLon(pixelX: Double, pixelY: Double, zoom: Int): Pair<Double, Dou
     val longitude = 360 * x
 
     return Pair(latitude, longitude)
-
 }
 
 /**
@@ -585,7 +584,7 @@ fun getReferenceCoordinate(path: LineString, targetDistance: Double, reverseLine
  * @return A bounding box object that contains coordinates for West/South/East/North edges
  * or min Lon / min Lat / max Lon / max Lat.
  */
-fun tileToBoundingBox(x: Int, y: Int, zoom: Double): BoundingBox {
+fun tileToBoundingBox(x: Int, y: Int, zoom: Int): BoundingBox {
     val boundingBox = BoundingBox()
     boundingBox.northLatitude = tileToLat(y, zoom)
     boundingBox.southLatitude = tileToLat(y + 1, zoom)
@@ -602,7 +601,7 @@ fun tileToBoundingBox(x: Int, y: Int, zoom: Double): BoundingBox {
  * Zoom level of the Slippy Tile
  * @return a latitude coordinate.
  */
-fun tileToLat(x: Int, zoom: Double): Double {
+fun tileToLat(x: Int, zoom: Int): Double {
     val n: Double = Math.PI - (2.0 * Math.PI * x) / 2.0.pow(zoom)
     return Math.toDegrees(atan(sinh(n)))
 }
@@ -615,7 +614,7 @@ fun tileToLat(x: Int, zoom: Double): Double {
  * Zoom level of the Slippy Tile.
  * @return a longitude coordinate.
  */
-fun tileToLon(y: Int, zoom: Double): Double {
+fun tileToLon(y: Int, zoom: Int): Double {
     return y / 2.0.pow(zoom) * 360.0 - 180
 }
